requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.0,<5.0
drf-spectacular>=0.27.0,<0.28
anthropic>=0.40.0
elevenlabs>=1.0.0
boto3>=1.34.0
celery>=5.3.0
channels>=4.0.0
channels-redis>=4.2.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-pptx>=0.6.23
stripe>=8.0.0
Pillow>=10.0.0
redis>=5.0.0
daphne>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
